# packages/api-client/openapi/workforce-api.yaml
openapi: 3.1.0
info:
  title: WORKFORCE API
  version: 1.1.0
  description: "Canonical Laravel API contract for Workforce ERP.


    User-facing business endpoints authenticate with Laravel Sanctum bearer

    tokens. `X-API-TOKEN` is reserved for explicitly scoped trusted

    server-to-server endpoints and is not a replacement for user authorization.

    "
servers:
  - url: http://localhost:8000
    description: Local Laravel API
...
paths:
...
  /api/v1/timesheets/clock-in:
    post:
      tags:
        - Timesheets
      operationId: clockIn
      summary: Clock an allowed employee in
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                employee_id:
                  type: integer
                  minimum: 1
      responses:
        "201":
          description: Clock-in created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesheetEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/timesheets/clock-out:
    post:
      tags:
        - Timesheets
      operationId: clockOut
      summary: Clock an allowed employee out of their active record
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                employee_id:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Clock-out completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesheetEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Laravel Sanctum personal access token
      description: Personal access token for explicitly non-browser API clients. Browser sign-in uses Sanctum session cookies and does not return a bearer token.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-TOKEN
      description: Environment-configured shared token used only by explicitly scoped trusted service endpoints.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorized for the operation/resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Route or resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current application state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
...
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required:
        - success
        - message
        - errors
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    TimesheetStatus:
      type: string
      enum:
        - present
        - absent
        - on-leave
        - half-day
        - pending
        - approved
        - rejected
    TimesheetEmployee:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    Timesheet:
      type: object
      additionalProperties: false
      required:
        - id
        - organization_id
        - employee_id
        - employee
        - date
        - clock_in
        - clock_out
        - total_hours
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
        organization_id:
          type: string
        employee_id:
          type: string
        employee:
          anyOf:
            - $ref: "#/components/schemas/TimesheetEmployee"
            - type: "null"
        date:
          type:
            - string
            - "null"
          format: date
        clock_in:
          type:
            - string
            - "null"
          format: date-time
        clock_out:
          type:
            - string
            - "null"
          format: date-time
        total_hours:
          type: number
          format: float
          minimum: 0
        status:
          $ref: "#/components/schemas/TimesheetStatus"
        created_at:
          type:
            - string
            - "null"
          format: date-time
        updated_at:
          type:
            - string
            - "null"
          format: date-time
...
    TimesheetEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          required:
            - data
          properties:
            data:
              $ref: "#/components/schemas/Timesheet"
...